requirements.txt
pandas
requests
tqdm
beautifulsoup4
yfinance
lxml
psycopg2-binary